Quality of life-improvements to combat the stupidity of the compiler …

+++ Source/main.cpp
@@ -17,7 +17,9 @@
   ******************************************************************************
   */
 /* USER CODE END Header */
-/* Includes ------------------------------------------------------------------*/
+/* Private includes -----------------------------------------------------------*/
+/* USER CODE BEGIN Includes */
+
 #include <stdint.h>
 #include <time.h>
 
@@ -33,7 +35,10 @@
 
 /* USER CODE END Includes */
 
+/* USER CODE END Includes */
+
 /* Private variables ---------------------------------------------------------*/
+/* USER CODE BEGIN PV */
 
 /* Handler for the ADC, DMA, ETH and SPI */
 ADC_HandleTypeDef hadc1;      
@@ -574,9 +579,8 @@ static void read_ADC(
    * 2048 % 3 = 2. Reducing the number of datapoints reduces the 
    * accuracy of the analysis, however prevents out-of-range error
    */
-  for(uint i = 0; i < DMA_BUFFER_LENGTH - 
-        NUM_HYDROPHONES; i++){
-    p_data_hyd_port[2 * i] = (float32_t)ADC1_converted_values[3 * i];
+  for(uint i = 0; i < DMA_BUFFER_LENGTH - NUM_HYDROPHONES; i++){
+    p_data_hyd_port[2 * i] = ADC1_converted_values[3 * i];
     p_data_hyd_port[(2 * i) + 1] = 0;
 
     p_data_hyd_starboard[2 * i] = (float32_t)ADC1_converted_values[(3 * i) + 1];
